GUI: Double buffering, API Update and rewrite of gfx utils
All the drawing is now done on a different buffer that is copied to the GRUB
allocated one when screen_update is called. This allows to avoid the
flickering that happens without this. New API update for framebuffer allows
to signal updating only parts of the screen for more optimizations. The
graphical libraries have also been rewritten to be faster and use the double
buffer